requirements.txt
altair
pandas
streamlit
matplotlib
seaborn
streamlit-aggrid
pyarrow
google-cloud
google-cloud-storage
numpy
